requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.65.0
requests>=2.28.0

# Optional dependencies for enhanced features
# opencv-python>=4.7.0  # For advanced image processing
# transformers>=4.30.0  # For LMM-based prompt enhancement